tapping on pokemon shows the image in top center
